web/requirements.txt
Django==5.2.6
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
requests==2.31.0
pytest==7.4.0
pytest-django==4.5.2
pytest-asyncio==0.21.0
requests-mock==1.11.0
freezegun==1.2.2
coverage==7.2.7
python-telegram-bot[job-queue]==20.7